Security and route guard.
Subjects added to UI.

+++ school/src/app/features/course/components/course/course.component.ts
@@ -2,7 +2,7 @@ import {AfterViewInit, Component, ElementRef, OnDestroy, OnInit, ViewChild} from
 import {fromEvent, merge, of, Subscription} from "rxjs";
 import {MatTable} from "@angular/material/table";
 import {MatDialog, MatDialogConfig} from "@angular/material/dialog";
-import {catchError, concatMap, debounceTime, distinctUntilChanged, tap} from "rxjs/operators";
+import {catchError, concatMap, debounceTime, distinctUntilChanged, switchMap, tap} from "rxjs/operators";
 import {CourseDetailComponent} from "../course-detail/course-detail.component";
 import {MatSort} from "@angular/material/sort";
 import {Course, CourseResolved} from "../../../../shared/models/course.model";
@@ -10,6 +10,7 @@ import {CourseService} from "../../../../shared/services/course.service";
 import {CourseDeleteConfirmationComponent} from "../course-delete-confirmation/course-delete-confirmation.component";
 import {MatPaginator} from "@angular/material/paginator";
 import {CourseDatasource} from "../../../../shared/services/course.datasource";
+import {AuthService} from "../../../../shared/services/auth.service";
 
 
 @Component({
@@ -31,14 +32,16 @@ export class CourseComponent implements OnInit, OnDestroy, AfterViewInit {
   @ViewChild('input', { static: false })
   input: ElementRef;
 
-
   dataSource: CourseDatasource;
-  $subscription = new Subscription();
   numCourses = 100;
 
   displayedColumns = ["id", "name",  "description", "roomName", "teacherLastName", "edit", "delete"]
 
+  $subscription = new Subscription();
+  isAuthenticated: boolean = false;
+
   constructor(private courseService: CourseService,
+              private authService: AuthService,
               private dialog: MatDialog) { }
 
   ngOnInit() {
@@ -49,36 +52,45 @@ export class CourseComponent implements OnInit, OnDestroy, AfterViewInit {
             count => this.numCourses = count
         )
     );
+
+    this.$subscription.add(
+      this.authService.authStatusChanges()
+        .subscribe(
+          authenticated => {
+            this.isAuthenticated = authenticated;
+          }
+        )
+    );
   }
 
   ngAfterViewInit() {
 
     this.$subscription.add(
-        this.sort.sortChange.subscribe(() => this.paginator.pageIndex = 0)
+      this.sort.sortChange.subscribe(() => this.paginator.pageIndex = 0)
     );
 
     this.$subscription.add(
-        fromEvent(this.input.nativeElement,'keyup')
-            .pipe(
-                debounceTime(350),
-                distinctUntilChanged(),
-                tap(() => {
-                  this.paginator.pageIndex = 0;
+      fromEvent(this.input.nativeElement,'keyup')
+          .pipe(
+              debounceTime(350),
+              distinctUntilChanged(),
+              tap(() => {
+                this.paginator.pageIndex = 0;
+                this.getCourses();
+              })
+          )
+          .subscribe()
+    );
+
+    this.$subscription.add(
+      merge(this.paginator.page, this.sort.sortChange)
+        .pipe(
+            tap(() => {
                   this.getCourses();
-                })
+                }
             )
-            .subscribe()
-    );
-
-    this.$subscription.add(
-        merge(this.paginator.page, this.sort.sortChange)
-            .pipe(
-                tap(() => {
-                      this.getCourses();
-                    }
-                )
-            )
-            .subscribe()
+        )
+        .subscribe()
     );
 
   }
@@ -103,29 +115,31 @@ export class CourseComponent implements OnInit, OnDestroy, AfterViewInit {
 
     const dialogRef =  this.dialog.open(CourseDetailComponent, dialogConfig);
 
-    dialogRef.afterClosed()
-      .pipe(
-        tap(formVal => {
-          if (formVal) {
-            formVal.id = course.id;
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          tap(formVal => {
+            if (formVal) {
+              formVal.id = course.id;
+            }
+          }),
+          concatMap(formVal => {
+            if (formVal) {
+              return this.courseService.updateCourse(formVal);
+            } else {
+              return of({});
+            }
+          }),
+          catchError(val => of(`error caught: ${val}`))
+        )
+        .subscribe(
+          returnVal => {
+            if (returnVal['result'] === 'success') {
+              this.getCourses();
+            }
           }
-        }),
-        concatMap(formVal => {
-          if (formVal) {
-            return this.courseService.updateCourse(formVal);
-          } else {
-            return of({});
-          }
-        }),
-        catchError(val => of(`error caught: ${val}`))
-      )
-      .subscribe(
-        returnVal => {
-          if (returnVal['result'] === 'success') {
-            this.getCourses();
-          }
-        }
-      )
+        )
+    );
   }
 
   createCourse() {
@@ -136,24 +150,26 @@ export class CourseComponent implements OnInit, OnDestroy, AfterViewInit {
 
     const dialogRef =  this.dialog.open(CourseDetailComponent, dialogConfig);
 
-    dialogRef.afterClosed()
-      .pipe(
-        concatMap(formVal => {
-          if (formVal) {
-            return this.courseService.createCourse(formVal);
-          } else {
-            return of({});
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          concatMap(formVal => {
+            if (formVal) {
+              return this.courseService.createCourse(formVal);
+            } else {
+              return of({});
+            }
+          }),
+          catchError(err => of(`error caught: ${err}`))
+        )
+        .subscribe(
+          returnVal => {
+            if (returnVal['result'] === 'success') {
+              this.getCourses();
+            }
           }
-        }),
-        catchError(err => of(`error caught: ${err}`))
-      )
-      .subscribe(
-        returnVal => {
-          if (returnVal['result'] === 'success') {
-            this.getCourses();
-          }
-        }
-      )
+        )
+    );
   }
 
 
@@ -169,19 +185,29 @@ export class CourseComponent implements OnInit, OnDestroy, AfterViewInit {
       }
     });
 
-    dialogRef.afterClosed().subscribe(dialogResult => {
-      if (dialogResult) {
-        this.courseService.deleteCourse(course.id)
-          .subscribe(
-            () => {
-              this.getCourses();
-            },
-            error => {
-              console.log('error in deletion of course ', error);
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          concatMap(dialogResult => {
+            if (dialogResult && dialogResult === true) {
+              return this.courseService.deleteCourse(course.id);
+            } else {
+              return of({'result': 'skipped'});
             }
+          }),
+          catchError(
+            err => { return of(`error caught: ${err}`) }
           )
-      }
-    });
+        )
+        .subscribe(
+          deletionResult => {
+            if (deletionResult && deletionResult.result && deletionResult.result === 'success') {
+              this.getCourses()
+            }
+          },
+          error => console.log('error in deletion of course ', error)
+        )
+    );
   }
 
   ngOnDestroy() {
