+++ school/src/app/features/room/components/room/room.component.ts
@@ -1,5 +1,5 @@
-import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
-import {Observable, of} from "rxjs";
+import {AfterViewInit, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {Observable, of, Subscription} from "rxjs";
 import {MatTable, MatTableDataSource} from "@angular/material/table";
 import {MatDialog, MatDialogConfig} from "@angular/material/dialog";
 import {catchError, concatMap, tap} from "rxjs/operators";
@@ -7,16 +7,15 @@ import {RoomDetailComponent} from "../room-detail/room-detail.component";
 import {MatSort} from "@angular/material/sort";
 import {RoomService} from "../../../../shared/services/room.service";
 import {Room} from "../../../../shared/models/room.model";
-import {ConfirmDialogComponent} from "../../../../shared/dialogs/confirm-dialog/confirm-dialog.component";
-import {TeacherDeleteConfirmationComponent} from "../../../teacher/components/teacher-delete-confirmation/teacher-delete-confirmation.component";
 import {RoomDeleteConfirmationComponent} from "../room-delete-confirmation/room-delete-confirmation.component";
+import {AuthService} from "../../../../shared/services/auth.service";
 
 @Component({
   selector: 'app-room',
   templateUrl: './room.component.html',
   styleUrls: ['./room.component.scss']
 })
-export class RoomComponent implements OnInit, AfterViewInit {
+export class RoomComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild(MatSort, { static: false })
   sort: MatSort;
@@ -28,30 +27,43 @@ export class RoomComponent implements OnInit, AfterViewInit {
   room$: Observable<Room[]>
   displayedColumns = ["id", "name", "capacity", "edit", "delete"]
 
+  $subscription = new Subscription();
+  isAuthenticated: boolean = false;
 
   constructor(private roomService: RoomService,
-              private dialog: MatDialog) {
-
-  }
+              private authService: AuthService,
+              private dialog: MatDialog) { }
 
   ngOnInit() {
     this.room$ = this.roomService.getRooms();
     this.getRooms();
+
+    this.$subscription.add(this.authService.authStatusChanges()
+      .subscribe(
+        authenticated => {
+          this.isAuthenticated = authenticated;
+        }
+      )
+    );
   }
 
   ngAfterViewInit() {
-    this.sort.sortChange.subscribe(() => {
-      this.dataSource.sortData(this.dataSource.data, this.sort);
-      this.matTable.renderRows();
-    });
+    this.$subscription.add(
+        this.sort.sortChange.subscribe(() => {
+        this.dataSource.sortData(this.dataSource.data, this.sort);
+        this.matTable.renderRows();
+      })
+    );
   }
 
   getRooms() {
-    this.room$.subscribe(
-      rooms => {
-        this.dataSource.data = rooms;
-      }
-    )
+    this.$subscription.add(
+      this.room$.subscribe(
+        rooms => {
+          this.dataSource.data = rooms;
+        }
+      )
+    );
   }
 
   searchRooms(search: string) {
@@ -67,29 +79,31 @@ export class RoomComponent implements OnInit, AfterViewInit {
 
     const dialogRef =  this.dialog.open(RoomDetailComponent, dialogConfig);
 
-    dialogRef.afterClosed()
-      .pipe(
-        tap(formVal => {
-          if(formVal) {
-            formVal.id = room.id;
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          tap(formVal => {
+            if(formVal) {
+              formVal.id = room.id;
+            }
+          }),
+          concatMap(formVal => {
+            if (formVal) {
+              return this.roomService.updateRoom(formVal);
+            } else {
+              return of({});
+            }
+          }),
+          catchError(val => of(`error caught: ${val}`))
+        )
+        .subscribe(
+          returnVal => {
+            if (returnVal['result'] === 'success') {
+              this.getRooms();
+            }
           }
-        }),
-        concatMap(formVal => {
-          if (formVal) {
-            return this.roomService.updateRoom(formVal);
-          } else {
-            return of({});
-          }
-        }),
-        catchError(val => of(`error caught: ${val}`))
-      )
-      .subscribe(
-        returnVal => {
-          if (returnVal['result'] === 'success') {
-            this.getRooms();
-          }
-        }
-      )
+        )
+    );
   }
 
   createRoom() {
@@ -100,24 +114,26 @@ export class RoomComponent implements OnInit, AfterViewInit {
 
     const dialogRef =  this.dialog.open(RoomDetailComponent, dialogConfig);
 
-    dialogRef.afterClosed()
-      .pipe(
-        concatMap(formVal => {
-          if (formVal) {
-            return this.roomService.createRoom(formVal);
-          } else {
-            return of({});
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          concatMap(formVal => {
+            if (formVal) {
+              return this.roomService.createRoom(formVal);
+            } else {
+              return of({});
+            }
+          }),
+          catchError(err => of(`error caught: ${err}`))
+        )
+        .subscribe(
+          returnVal => {
+            if (returnVal['result'] === 'success') {
+              this.getRooms();
+            }
           }
-        }),
-        catchError(err => of(`error caught: ${err}`))
-      )
-      .subscribe(
-        returnVal => {
-          if (returnVal['result'] === 'success') {
-            this.getRooms();
-          }
-        }
-      )
+        )
+    );
   }
 
   deleteRoom(room: Room) {
@@ -132,17 +148,34 @@ export class RoomComponent implements OnInit, AfterViewInit {
       }
     });
 
-
-    dialogRef.afterClosed().subscribe(dialogResult => {
-      if (dialogResult) {
-        this.roomService.deleteRoom(room.id)
-          .subscribe(
-            () => {
-              this.getRooms();
+    this.$subscription.add(
+      dialogRef.afterClosed()
+        .pipe(
+          concatMap(dialogResult => {
+            if (dialogResult && dialogResult === true) {
+              return this.roomService.deleteRoom(room.id);
+            } else {
+              return of({'result': 'skipped'});
             }
+          }),
+          catchError(
+            err => { return of(`error caught: ${err}`) }
           )
-      }
-    });
+        )
+        .subscribe(
+          deletionResult => {
+            if (deletionResult && deletionResult.result && deletionResult.result === 'success') {
+              this.getRooms()
+            }
+          },
+          error => console.log('error in deletion of room ', error)
+        )
+    );
+
+  }
+
+  ngOnDestroy() {
+    this.$subscription.unsubscribe();
   }
 
 }
